hotel_revenue_optimization_ui/requirements.txt
flask==2.3.3
flask-wtf==1.1.1
boto3>=1.34.0,<2.0.0
botocore>=1.34.0,<2.0.0
markdown==3.5
gunicorn==23.0.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.3
pyjwt==2.8.0
cryptography==42.0.8
